feat: add realtime update to Members
