+++ src/components/TaskPage/TaskPageContainer.jsx
@@ -3,13 +3,11 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import TaskPage from './TaskPage';
 import { getTask, getUsersWithTask, clearUserTasks } from '../../redux/reducers/taskPageReducer';
-import { getMembers } from '../../redux/reducers/membersReducer';
 
 class TaskPageContainer extends React.Component {
   componentDidMount() {
-    const { taskId, getTask, getMembers, getUsersWithTask, clearUserTasks } = this.props;
+    const { taskId, getTask, getUsersWithTask, clearUserTasks } = this.props;
     getTask(taskId);
-    getMembers();
     getUsersWithTask(taskId);
     clearUserTasks();
   }
@@ -37,4 +35,4 @@ TaskPageContainer.defaultProps = {
   taskId: '',
 };
 
-export default connect(mapStateToProps, { getTask, getMembers, getUsersWithTask, clearUserTasks })(TaskPageContainer);
+export default connect(mapStateToProps, { getTask, getUsersWithTask, clearUserTasks })(TaskPageContainer);
